Fix conflict between widget + shortcode js, fix other widget issues.
The enqueue slugs for both scripts were accidentally set to the same
thing, the page builder plugin includes widget code on edit post pages
so the widget script was replacing the shortcode script.

While fixing this ran into a few issues with how the widget edit preview
was working, and how widgets handled the show property. Should be
resolved now.